applied some of requested changes

+++ Explorer/Assets/DCL/PerformanceAndDiagnostics/PerformanceBudgeting/Memory/StandaloneSystemMemory.cs
@@ -4,6 +4,6 @@ namespace DCL.PerformanceBudgeting
 {
     public class StandaloneSystemMemory : ISystemMemory
     {
-        public long TotalSizeInMB => SystemInfo.systemMemorySize;
+        public ulong TotalSizeInMB => (ulong)SystemInfo.systemMemorySize;
     }
 }
